Change the API to use a channel of events.
Instead of having callback functions, the library publishes events
to an event channel. The caller can type switch on the event and then
take the appropriate action.

I ported the callback code over to event code basically 1-to-1. One
issue with this is some of the events fire as part of Dial() which
is fairly awkward and would mean we would deadlock if the channel was
unbuffered. To work around this for now we use a buffered channel. I
think we should evaluate if those events that happen in Dial()
actually have any value.

Update the player demo script to use the new api.